pyproject.toml
[tool.poetry]
name = "images"
version = "0.1.0"
description = ""
authors = ["hkorzeniewski <56238495+hkorzeniewski@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
environ = "^1.0"
django-environ = "^0.10.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
django-filter = "^22.1"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
bandit = "^1.7.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
